Cybersecurity Incident Report: Network Traffic
This lab project shows a scenario for a situation with a port issue with a company website that needed help investigating potential malicious network activity traffic to detect a potential threat within the protocol. This incident report breaks down the methods used to find the potential threat activity that was detected within this process in this written document.