SecurityIncidentReport for "Yummyrecipesforme.com" CyberHack
In this incident, this shows an issue where a website got hacked and it was replaced with a fake website. This website “yummyrecipesfor.com” was compromised by a malicious cyber-attacker to exploit the security vulnerabilities and utilize any means to attack the security framework to take advantage of anything necessary for criminal means. This documentation shows the steps taken to combat against the issue.